♻️ refactor(SCRUM-265): PortfolioStepper → StepperWithText로 컴포넌트명 변경 & NewPortfolioPage 파일명 index.tsx로 수정 & 포트폴리오 관련 컴포넌트들 폴더 이동
