+++ src/App.tsx
@@ -2,7 +2,7 @@ import Header from '@/components/layout/Header';
 import RecruitPage from '@/components/pages/recruit';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import PortfolioPage from '@/components/pages/portfolio';
-import NewPortfolioPage from '@/components/pages/portfolio/new/NewPortfolioPage';
+import NewPortfolioPage from '@/components/pages/portfolio/new';
 
 const App: React.FC = () => {
   return (
